refactor: 채키지 정리
