+++ src/main/java/janggi/game/JanggiBoard.java
@@ -1,9 +1,10 @@
-package janggi.board;
+package janggi.game;
 
 import janggi.piece.Piece;
 import janggi.piece.Pieces;
 import janggi.setting.CampType;
 import janggi.setting.PieceAssignType;
+import janggi.setting.PieceAssigner;
 import janggi.value.Position;
 import java.util.ArrayList;
 import java.util.Comparator;
