+++ src/main/java/janggi/game/JanggiGame.java
@@ -1,6 +1,5 @@
 package janggi.game;
 
-import janggi.board.JanggiBoard;
 import janggi.setting.CampType;
 import janggi.setting.PieceAssignType;
 import janggi.value.Position;
@@ -46,8 +45,7 @@ public class JanggiGame {
         PieceAssignType choAnswer = readPieceAssignType(CampType.CHO);
         PieceAssignType hanAnswer = readPieceAssignType(CampType.HAN);
         JanggiBoard janggiBoard = new JanggiBoard(choAnswer, hanAnswer);
-        outputView.writeScore(janggiBoard.getScore(CampType.CHO), janggiBoard.getScore(CampType.HAN));
-        outputView.writeJanggiBoard(janggiBoard.getPieces(CampType.CHO), janggiBoard.getPieces(CampType.HAN));
+        printJanggiBoardState(janggiBoard);
         return janggiBoard;
     }
 
@@ -58,8 +56,7 @@ public class JanggiGame {
                 Position movedPiecePosition = inputView.readMovedPiecePosition();
                 Position destination = inputView.readDestinationPosition();
                 janggiBoard.movePiece(campType, movedPiecePosition, destination);
-                outputView.writeScore(janggiBoard.getScore(CampType.CHO), janggiBoard.getScore(CampType.HAN));
-                outputView.writeJanggiBoard(janggiBoard.getPieces(CampType.CHO), janggiBoard.getPieces(CampType.HAN));
+                printJanggiBoardState(janggiBoard);
                 return;
             } catch (IllegalArgumentException exception) {
                 outputView.printExceptionMessage(exception.getMessage());
@@ -85,6 +82,11 @@ public class JanggiGame {
                 outputView.printExceptionMessage(exception.getMessage());
             }
         }
+    }
+
+    private void printJanggiBoardState(JanggiBoard board) {
+        outputView.writeScore(board.getScore(CampType.CHO), board.getScore(CampType.HAN));
+        outputView.writeJanggiBoard(board.getPieces(CampType.CHO), board.getPieces(CampType.HAN));
     }
 
     private void printGameResult(JanggiBoard board) {
